pyproject.toml
[tool.poetry]
name = "mslearn-ai-language"
version = "0.1.0"
description = "Develop natural language processing solutions with Azure AI Services"
authors = ["Sérgio Pereira da Silva <sergio.pds@outlook.pt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
azure-ai-language-questionanswering = "^1.1.0"
azure-ai-textanalytics = "5.3.0"
python-dotenv = "^1.0.0"
azure-ai-language-conversations = "^1.1.0"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
